refactor: ♻️ rename plugins to preset-plugins
