+++ packages/NarrativeTextEditor/src/preset-plugins/core/moveSelectionByOffset.ts
