+++ packages/NarrativeTextEditor/src/preset-plugins/utils/exit-break.ts
