+++ packages/NarrativeTextEditor/src/preset-plugins/utils/soft-break.ts
