+++ packages/NarrativeTextEditor/src/preset-plugins/variable/VariableCombobox/getVariableOnSelectItem.ts
