+++ packages/NarrativeTextEditor/src/preset-plugins/variable/VariableElement/index.tsx
