+++ packages/NarrativeTextEditor/src/preset-plugins/variable/VariableInputElement/index.tsx
