+++ packages/NarrativeTextEditor/src/preset-plugins/variable/constants.ts
