+++ packages/NarrativeTextEditor/src/preset-plugins/variable/createVariablePlugin.ts
