+++ packages/NarrativeTextEditor/src/preset-plugins/variable/handlers/index.ts
