+++ packages/NarrativeTextEditor/src/preset-plugins/variable/handlers/variableOnKeyDownHandler.ts
