+++ packages/NarrativeTextEditor/src/preset-plugins/variable/index.tsx
