+++ packages/NarrativeTextEditor/src/preset-plugins/variable/queries/findVariableInput.ts
