+++ packages/NarrativeTextEditor/src/preset-plugins/variable/queries/index.ts
