+++ packages/NarrativeTextEditor/src/preset-plugins/variable/queries/isNodeVariableInput.ts
