+++ packages/NarrativeTextEditor/src/preset-plugins/variable/queries/isSelectionInVariableInput.ts
