+++ packages/NarrativeTextEditor/src/preset-plugins/variable/transforms/removeVariableInput.ts
