+++ packages/NarrativeTextEditor/src/preset-plugins/variable/types.ts
