+++ packages/NarrativeTextEditor/src/preset-plugins/variable/withVariable.ts
